refactor: Improve code readability and error handling in calcularDataInicioFim.py and tratarPlanilhaBilhetesVendidos.py
